+++ data/conditions.ts
@@ -364,7 +364,7 @@ export const Conditions: {[k: string]: ConditionData} = {
 			if (data.source.hasAbility('adaptability') && this.gen >= 6) {
 				data.moveData.stab = 2;
 			}
-			const hitMove = new this.dex.Data.Move(data.moveData) as Move as ActiveMove;
+			const hitMove = new this.dex.Move(data.moveData) as ActiveMove;
 
 			this.trySpreadMoveHit([target], data.source, hitMove);
 		},
