+++ data/mods/letsgo/moves.ts
@@ -50,7 +50,7 @@ export const Moves: {[k: string]: ModdedMoveData} = {
 			}
 			let randomMove = '';
 			if (moves.length) {
-				moves.sort((a, b) => a.num! - b.num!);
+				moves.sort((a, b) => a.num - b.num);
 				randomMove = this.sample(moves).id;
 			}
 			if (!randomMove) return false;
