+++ sim/dex.ts
@@ -32,6 +32,12 @@ import * as fs from 'fs';
 import * as path from 'path';
 
 import * as Data from './dex-data';
+import {DataCondition} from './dex-conditions';
+import {DataMove} from './dex-moves';
+import {DataItem} from './dex-items';
+import {DataAbility} from './dex-abilities';
+import {Species} from './dex-species';
+import {DataFormat, RuleTable, mergeFormatLists, ComplexBan, ComplexTeamBan} from './dex-formats';
 import {PRNG, PRNGSeed} from './prng';
 import {Utils} from '../lib/utils';
 
@@ -67,13 +73,13 @@ const DATA_FILES = {
 	TypeChart: 'typechart',
 };
 
-const nullEffect: Condition = new Data.Condition({name: '', exists: false});
+const nullEffect: Condition = new DataCondition({name: '', exists: false});
 
 interface DexTableData {
 	Abilities: DexTable<Ability>;
 	Aliases: {[id: string]: string};
 	Formats: DexTable<Format>;
-	FormatsData: DexTable<ModdedSpeciesFormatsData>;
+	FormatsData: DexTable<import('./dex-species').ModdedSpeciesFormatsData>;
 	Items: DexTable<Item>;
 	Learnsets: DexTable<LearnsetData>;
 	Moves: DexTable<Move>;
@@ -93,84 +99,34 @@ interface TextTableData {
 
 export const toID = Data.toID;
 
-// function for merging the two lists
-function mergeFormats(main: FormatList, custom: FormatList | undefined): FormatList {
-	// interface for the builder.
-	interface FormatSection {
-		section: string;
-		column?: number;
-		formats: FormatData[];
-	}
-
-	// result that is return and makes the actual list for formats.
-	const result: FormatList = [];
-
-	// used as a intermediary to build the final list.
-	const build: FormatSection[] = [];
-
-	// used to track current section to keep formats under their sections.
-	let current: FormatSection | undefined = {section: "", formats: []};
-
-	// populates the original sections and formats easily
-	// there should be no repeat sections at this point.
-	for (const element of main) {
-		if (element.section) {
-			current = {section: element.section, column: element.column, formats: []};
-			build.push(current);
-		} else if ((element as FormatData).name) {
-			current.formats.push((element as FormatData));
-		}
-	}
-
-	// merges the second list the hard way. Accounts for repeats.
-	if (custom !== undefined) {
-		for (const element of custom) {
-			// finds the section and makes it if it doesn't exist.
-			if (element.section) {
-				current = build.find(e => e.section === element.section);
-
-				// if it's new it makes a new entry.
-				if (current === undefined) {
-					current = {section: element.section, column: element.column, formats: []};
-					build.push(current);
-				}
-			} else if ((element as FormatData).name) { // otherwise, adds the element to its section.
-				current.formats.push(element as FormatData);
-			}
-		}
-	}
-
-	// builds the final result.
-	for (const element of build) {
-		// adds the section to the list.
-		result.push({section: element.section, column: element.column}, ...element.formats);
-	}
-
-	return result;
-}
-
 export class ModdedDex {
-	readonly Data: typeof Data;
-	readonly ModdedDex: typeof ModdedDex;
+	readonly Data = Data;
+	readonly Condition = DataCondition;
+	readonly Ability = DataAbility;
+	readonly Item = DataItem;
+	readonly Move = DataMove;
+	readonly Species = Species;
+	readonly Format = DataFormat;
+	readonly ModdedDex = ModdedDex;
 
-	readonly name: string;
+	readonly name = "[ModdedDex]";
 	readonly isBase: boolean;
 	readonly currentMod: string;
 
-	readonly toID: (text: any) => ID;
+	readonly toID = Data.toID;
 
-	readonly abilityCache: Map<ID, Ability>;
-	readonly effectCache: Map<ID, Effect | Move>;
-	readonly itemCache: Map<ID, Item>;
-	readonly learnsetCache: Map<ID, LearnsetData>;
-	readonly moveCache: Map<ID, Move>;
-	readonly speciesCache: Map<ID, Species>;
-	readonly natureCache: Map<ID, Nature>;
-	readonly typeCache: Map<string, TypeInfo>;
+	readonly abilityCache = new Map<ID, Ability>();
+	readonly effectCache = new Map<ID, Effect | Move>();
+	readonly itemCache = new Map<ID, Item>();
+	readonly learnsetCache = new Map<ID, LearnsetData>();
+	readonly moveCache = new Map<ID, Move>();
+	readonly speciesCache = new Map<ID, Species>();
+	readonly natureCache = new Map<ID, Nature>();
+	readonly typeCache = new Map<string, TypeInfo>();
 
-	gen: number;
-	parentMod: string;
-	modsLoaded: boolean;
+	gen = 0;
+	parentMod = '';
+	modsLoaded = false;
 
 	dataCache: DexTableData | null;
 	textCache: TextTableData | null;
@@ -179,27 +135,8 @@ export class ModdedDex {
 	deepClone = Utils.deepClone;
 
 	constructor(mod = 'base', isOriginal = false) {
-		this.ModdedDex = ModdedDex;
-		this.Data = Data;
-
-		this.name = "[ModdedDex]";
 		this.isBase = (mod === 'base');
 		this.currentMod = mod;
-
-		this.toID = toID;
-
-		this.abilityCache = new Map();
-		this.effectCache = new Map();
-		this.itemCache = new Map();
-		this.moveCache = new Map();
-		this.learnsetCache = new Map();
-		this.speciesCache = new Map();
-		this.natureCache = new Map();
-		this.typeCache = new Map();
-
-		this.gen = 0;
-		this.parentMod = '';
-		this.modsLoaded = false;
 
 		this.dataCache = null;
 		this.textCache = null;
@@ -366,7 +303,7 @@ export class ModdedDex {
 			if (this.data.FormatsData.hasOwnProperty(id)) {
 				// special event ID, like Rockruff-Dusk
 				const baseId = toID(this.data.Aliases[id]);
-				species = new Data.Species({name}, this.data.Pokedex[baseId], this.data.FormatsData[id]);
+				species = new Species({name}, this.data.Pokedex[baseId], this.data.FormatsData[id]);
 				species.name = id;
 				species.name = id;
 				species.id = id;
@@ -376,7 +313,7 @@ export class ModdedDex {
 				if (species.cosmeticFormes) {
 					for (const forme of species.cosmeticFormes) {
 						if (toID(forme) === id) {
-							species = new Data.Species(species, {
+							species = new Species(species, {
 								name: forme,
 								id,
 								forme: forme.slice(species.name.length + 1),
@@ -427,7 +364,7 @@ export class ModdedDex {
 			}
 		}
 		if (id && this.data.Pokedex.hasOwnProperty(id)) {
-			species = new Data.Species({name}, this.data.Pokedex[id], this.data.FormatsData[id]);
+			species = new Species({name}, this.data.Pokedex[id], this.data.FormatsData[id]);
 			// Inherit any statuses from the base species (Arceus, Silvally).
 			const baseSpeciesStatuses = this.data.Conditions[toID(species.baseSpecies)];
 			if (baseSpeciesStatuses !== undefined) {
@@ -474,7 +411,7 @@ export class ModdedDex {
 				(!['Ditto', 'Undiscovered'].includes(species.eggGroups[0]) && !species.prevo && species.name !== 'Manaphy');
 			if (this.gen === 1) species.bst -= species.baseStats.spd;
 		} else {
-			species = new Data.Species({
+			species = new Species({
 				id, name, exists: false, tier: 'Illegal', doublesTier: 'Illegal', isNonstandard: 'Custom',
 			});
 		}
@@ -542,12 +479,12 @@ export class ModdedDex {
 		if (id && this.data.Moves.hasOwnProperty(id)) {
 			const moveData = this.data.Moves[id];
 			const moveTextData = this.getDescs('Moves', id, moveData);
-			move = new Data.Move({name}, moveData, moveTextData);
+			move = new DataMove({name}, moveData, moveTextData);
 			if (move.gen > this.gen) {
 				(move as any).isNonstandard = 'Future';
 			}
 		} else {
-			move = new Data.Move({id, name, exists: false});
+			move = new DataMove({id, name, exists: false});
 		}
 		if (move.exists) this.moveCache.set(id, move);
 		return move;
@@ -605,21 +542,21 @@ export class ModdedDex {
 
 		let found;
 		if (this.data.Formats.hasOwnProperty(id)) {
-			effect = new Data.Format({name: id}, this.data.Formats[id]);
+			effect = new DataFormat({name: id}, this.data.Formats[id]);
 		} else if (this.data.Conditions.hasOwnProperty(id)) {
-			effect = new Data.Condition({name: id}, this.data.Conditions[id]);
+			effect = new DataCondition({name: id}, this.data.Conditions[id]);
 		} else if (
 			(this.data.Moves.hasOwnProperty(id) && (found = this.data.Moves[id]).condition) ||
 			(this.data.Abilities.hasOwnProperty(id) && (found = this.data.Abilities[id]).condition) ||
 			(this.data.Items.hasOwnProperty(id) && (found = this.data.Items[id]).condition)
 		) {
-			effect = new Data.Condition({name: found.name || id}, found.condition);
+			effect = new DataCondition({name: found.name || id}, found.condition);
 		} else if (id === 'recoil') {
-			effect = new Data.Condition({id, name: 'Recoil', effectType: 'Recoil'});
+			effect = new DataCondition({id, name: 'Recoil', effectType: 'Recoil'});
 		} else if (id === 'drain') {
-			effect = new Data.Condition({id, name: 'Drain', effectType: 'Drain'});
+			effect = new DataCondition({id, name: 'Drain', effectType: 'Drain'});
 		} else {
-			effect = new Data.Condition({id, name: id, exists: false});
+			effect = new DataCondition({id, name: id, exists: false});
 		}
 
 		this.effectCache.set(id, effect);
@@ -680,9 +617,9 @@ export class ModdedDex {
 		}
 		let effect;
 		if (this.data.Formats.hasOwnProperty(id)) {
-			effect = new Data.Format({name}, this.data.Formats[id], supplementaryAttributes);
+			effect = new DataFormat({name}, this.data.Formats[id], supplementaryAttributes);
 		} else {
-			effect = new Data.Format({id, name, exists: false});
+			effect = new DataFormat({id, name, exists: false});
 		}
 		return effect;
 	}
@@ -709,7 +646,7 @@ export class ModdedDex {
 		if (id && this.data.Items.hasOwnProperty(id)) {
 			const itemData = this.data.Items[id];
 			const itemTextData = this.getDescs('Items', id, itemData);
-			item = new Data.Item({name}, itemData, itemTextData);
+			item = new DataItem({name}, itemData, itemTextData);
 			if (item.gen > this.gen) {
 				(item as any).isNonstandard = 'Future';
 			}
@@ -718,7 +655,7 @@ export class ModdedDex {
 				(item as any).isNonstandard = 'Past';
 			}
 		} else {
-			item = new Data.Item({id, name, exists: false});
+			item = new DataItem({id, name, exists: false});
 		}
 
 		if (item.exists) this.itemCache.set(id, item);
@@ -741,7 +678,7 @@ export class ModdedDex {
 		if (id && this.data.Abilities.hasOwnProperty(id)) {
 			const abilityData = this.data.Abilities[id];
 			const abilityTextData = this.getDescs('Abilities', id, abilityData);
-			ability = new Data.Ability({name}, abilityData, abilityTextData);
+			ability = new DataAbility({name}, abilityData, abilityTextData);
 			if (ability.gen > this.gen) {
 				(ability as any).isNonstandard = 'Future';
 			}
@@ -752,7 +689,7 @@ export class ModdedDex {
 				(ability as any).isNonstandard = null;
 			}
 		} else {
-			ability = new Data.Ability({id, name, exists: false});
+			ability = new DataAbility({id, name, exists: false});
 		}
 
 		if (ability.exists) this.abilityCache.set(id, ability);
@@ -839,13 +776,13 @@ export class ModdedDex {
 		}
 	}
 
-	getRuleTable(format: Format, depth = 1, repeals?: Map<string, number>): Data.RuleTable {
+	getRuleTable(format: Format, depth = 1, repeals?: Map<string, number>): RuleTable {
 		if (format.ruleTable && !repeals) return format.ruleTable;
 		if (depth === 1 && dexes[format.mod || 'base'] !== this) {
 			// throw new Error(`${format.mod} ${this.currentMod}`);
 			return this.mod(format.mod).getRuleTable(format, depth + 1);
 		}
-		const ruleTable = new Data.RuleTable();
+		const ruleTable = new RuleTable();
 
 		const ruleset = format.ruleset.slice();
 		for (const ban of format.banlist) {
@@ -885,10 +822,10 @@ export class ModdedDex {
 
 			if (typeof ruleSpec !== 'string') {
 				if (ruleSpec[0] === 'complexTeamBan') {
-					const complexTeamBan: Data.ComplexTeamBan = ruleSpec.slice(1) as Data.ComplexTeamBan;
+					const complexTeamBan: ComplexTeamBan = ruleSpec.slice(1) as ComplexTeamBan;
 					ruleTable.addComplexTeamBan(complexTeamBan[0], complexTeamBan[1], complexTeamBan[2], complexTeamBan[3]);
 				} else if (ruleSpec[0] === 'complexBan') {
-					const complexBan: Data.ComplexBan = ruleSpec.slice(1) as Data.ComplexBan;
+					const complexBan: ComplexBan = ruleSpec.slice(1) as ComplexBan;
 					ruleTable.addComplexBan(complexBan[0], complexBan[1], complexBan[2], complexBan[3]);
 				} else {
 					throw new Error(`Unrecognized rule spec ${ruleSpec}`);
@@ -1554,7 +1491,7 @@ export class ModdedDex {
 			}
 		}
 		try {
-			Formats = mergeFormats(require(MAIN_FORMATS).Formats, customFormats);
+			Formats = mergeFormatLists(require(MAIN_FORMATS).Formats, customFormats);
 		} catch (e) {
 			if (e.code !== 'MODULE_NOT_FOUND' && e.code !== 'ENOENT') {
 				throw e;
@@ -1596,3 +1533,13 @@ dexes['base'] = new ModdedDex(undefined, true);
 dexes[BASE_MOD] = dexes['base'];
 
 export const Dex = dexes['base'];
+export namespace Dex {
+	export type Species = import('./dex-species').Species;
+	export type Item = import('./dex-items').Item;
+	export type Move = import('./dex-moves').Move;
+	export type Ability = import('./dex-abilities').Ability;
+
+	export type HitEffect = import('./dex-moves').HitEffect;
+	export type SecondaryEffect = import('./dex-moves').SecondaryEffect;
+	export type RuleTable = import('./dex-formats').RuleTable;
+}
